Consider plugged in as charging
There are scenarios when the device is neither charging nor fully
charged but has a charger connected, for example battery preservation
features that disable or delay charging under certain conditions.

This considers a plugged-in charger as "charging" as a charger typically
provides enough power to at least maintain the battery level.

Signed-off-by: Daniel Höxtermann <daniel@hxtm.dev>